NEM-196 / bug : sprint1_merge_테스트(충돌 오류 수정)
- Product, Member merge 후 발생한 충돌 해결
