+++ server/src/docs/asciidoc/product.adoc
