--- server/src/main/java/com/noterror/app/api/README.md
@@ -1,3 +1,0 @@
-# API
-
-핵심 관심사들을 해당 패키지에 정리합니다.
