+++ server/src/main/java/com/noterror/app/api/domain/product/dto/ProductPatchDto.java
