+++ server/src/main/java/com/noterror/app/api/domain/product/dto/ProductResponseDto.java
@@ -1,6 +1,6 @@
 package com.noterror.app.api.domain.product.dto;
 
-import com.noterror.app.api.domain.product.entity.Product;
+import com.noterror.app.api.domain.entity.Product;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
